took Card class out of Deck, closes #23
