+++ CardGames/pile.py
@@ -1,5 +1,3 @@
-from deck import Card
-
 class Pile:
 
     def __init__(self):
